+++ zebra-state/src/service/finalized_state/zebra_db/rollback.rs
@@ -558,12 +558,24 @@ fn rebuild_history_tree_from_upgrade_activation(
         .expect("current network upgrade must have an activation height");
 
     let (block, sapling_root, orchard_root) = history_rebuild_inputs_at_height(db, start_height)?;
-    let mut history_tree = HistoryTree::from_block(network, block, &sapling_root, &orchard_root)?;
+    let mut history_tree = HistoryTree::from_block(
+        network,
+        block,
+        &sapling_root,
+        &orchard_root,
+        &Default::default(),
+    )?;
 
     for height in ((start_height.0 + 1)..=target_height.0).map(Height) {
         let (block, sapling_root, orchard_root) = history_rebuild_inputs_at_height(db, height)?;
 
-        history_tree.push(network, block, &sapling_root, &orchard_root)?;
+        history_tree.push(
+            network,
+            block,
+            &sapling_root,
+            &orchard_root,
+            &Default::default(),
+        )?;
     }
 
     Ok(history_tree)
@@ -614,7 +626,13 @@ fn rebuild_treestate_to_height(
 
         let sapling_root = note_commitment_trees.sapling.root();
         let orchard_root = note_commitment_trees.orchard.root();
-        history_tree.push(network, block, &sapling_root, &orchard_root)?;
+        history_tree.push(
+            network,
+            block,
+            &sapling_root,
+            &orchard_root,
+            &Default::default(),
+        )?;
     }
 
     Ok(RebuiltTreestate {