+++ zebra-state/src/service/non_finalized_state/tests/vectors.rs
@@ -658,6 +658,7 @@ fn history_tree_is_updated_for_network_upgrade(
         activation_block.clone(),
         &chain.sapling_note_commitment_tree_for_tip().root(),
         &chain.orchard_note_commitment_tree_for_tip().root(),
+        &Default::default(),
     )
     .unwrap();
 
@@ -741,6 +742,7 @@ fn commitment_is_validated_for_network_upgrade(network: Network, network_upgrade
         activation_block.clone(),
         &chain.sapling_note_commitment_tree_for_tip().root(),
         &chain.orchard_note_commitment_tree_for_tip().root(),
+        &Default::default(),
     )
     .unwrap();